InvertedPendulum/pyproject.toml
[project]
name = "invertedpendulum"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "control==0.9.4",
    "dm-control==1.0.17",
    "matplotlib==3.8.2",
    "moviepy==1.0.3",
    "mujoco==3.1.4",
    "numpy==1.26.2",
    "opencv-python==4.8.1.78",
    "pandas==2.2.2",
    "pillow==10.1.0",
    "scipy==1.11.4",
    "tensorboard==2.16.2",
    "tensorflow>=2.16.2",
    "torch==2.2.2",
    "torchaudio==2.2.2",
    "torchvision==0.17.2",
]
#dependencies = [
#    "numpy>=2.1.3",
#    "tensorflow>=2.19.0",
#]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
